[Chore #163057898] Separate Test File.
